backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
transformers==4.35.2
torch==2.9.1
scikit-learn==1.3.2
pillow==10.1.0
fastapi-users[sqlalchemy]==12.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.0.0
